+++ oodar.c
@@ -19,6 +19,8 @@
 #include "sec/cap/cap_alloc.c"
 #include "sec/cap/cap_time.c"
 #include "sec/cap/cap_ffi.c"
+#include "sec/cap/cap_bridge_emitted.c"
+#include "sec/cap/cap_ocap_bridge.c"
 #include "sec/cap/cap_check_zig_fallback.c"
 #include "sec/landlock/landlock/landlock.c"
 #include "sec/landlock/landlock/landlock_restrict.c"